\name{dumpDSA}
\alias{dumpDSA}
\title{dumpDSA}
\description{
\code{dumpDSA} is used to create an XML file containing
visualization information for a partDSA object.
}
\usage{
dumpDSA(x, file=stdout())
}
\arguments{
\item{x}{\code{partDSA} or \code{dsa} object to be dumped.}
\item{file}{file to write data to.}
}

\keyword{utilities}
