\name{loglevel}
\alias{loglevel}
\title{Set Logging Level}
\description{
\code{loglevel} is used to set the logging level for the DSA package.
}
\usage{
loglevel(lev)
}
\arguments{
\item{lev}{The default level to use.  The level can be specified
as \code{error}, \code{warn}, \code{info}, \code{debug},
or \code{verbose}.
}
}
\examples{
loglevel(debug)
}
\keyword{utilities}
