\name{predict.dsa}
\alias{predict.dsa}
\title{predict.dsa}
\description{
Predicted values based on the DSA algorithm.
}
\usage{
\method{predict}{dsa}(object, newdata1, \dots)
}
\arguments{
\item{object}{a \code{partDSA} or \code{dsa} object.}
\item{newdata1}{a data frame in which to look for variables with
which to predict.}
\item{\dots}{unused.}
}
\keyword{utilities}
